respect requested interval
