+++ utils.js
@@ -2,6 +2,7 @@ const cloneDeep = require('lodash/cloneDeep');
 const filter = require('lodash/filter');
 const partition = require('lodash/partition');
 const keyBy = require('lodash/keyBy');
+const orderBy = require('lodash/orderBy');
 const Big = require('big.js');
 const { format, eachDayOfInterval, parse, isAfter, isBefore } = require('date-fns');
 
@@ -94,7 +95,69 @@ function getPreviousValue(arr, i) {
   }
 }
 
+function calcSalesDataFIFO(activities) {
+  activities = orderBy(cloneDeep(activities), 'date', 'asc');
+
+  // we will make changes to sales that should be reflected in the activities collection
+  const sales = activities.filter((a) => ['Sell', 'TransferOut'].includes(a.type));
+
+  // we clone purchases again as the changes we will make below are temporary and should not be returned
+  const purchases = cloneDeep(activities.filter((a) => ['Buy', 'TransferIn'].includes(a.type)));
+
+  /**
+   * Alright, in here we will loop through sales
+   * and then start with the first purchase to see how much we paid originally
+   * for the shares we are selling here
+   */
+  sales.forEach((sale) => {
+    const { shares } = sale;
+    sale.buyAmount = 0;
+
+    // alright let's do the FIFO magic in here
+    function calc(sellShares) {
+      const purchase = purchases[0];
+      if (!purchase) {
+        return;
+      }
+
+      // get the shares and the price of the first purchase
+      const buyShares = purchase.shares;
+      const buyPrice = purchase.price;
+
+      // check how many shares were actually sold in this purchase
+      const sharesSoldInThisPurchase = Math.min(buyShares, sellShares);
+
+      // these are the remaining shares, in case we sold more shares than we bought in this purchase
+      // this means there are more purchases where we need to calc the remaining shares against
+      const remainingShares = +Big(buyShares).minus(Big(sellShares));
+
+      // store the remaining shares on the purchase - for the next sale that will happen
+      purchase.shares = remainingShares;
+
+      // store the amount of this purchase to our sale
+      sale.buyAmount += buyPrice * sharesSoldInThisPurchase;
+
+      // if no shares are left (all of this purchase was sold in the sale), remove the purchase
+      if (remainingShares <= 0) {
+        purchases.shift();
+      }
+
+      // if the remaining shares are negative (ie we sold more than we bought in this purchase)
+      // restart this calculation with the remaining shares as "sellShares"
+      if (remainingShares < 0) {
+        calc(Math.abs(remainingShares));
+      }
+    }
+
+    // kick FIFO off
+    calc(shares);
+  });
+
+  return activities;
+}
+
 module.exports = {
+  calcSalesDataFIFO,
   applySplitMultiplier,
   getDateArr,
   normalizeQuotes,
